+++ src/main/scala/controllers/SearchController.scala
@@ -25,12 +25,16 @@ import config.AppConfig
 import models.{CompaniesHouseId, ReportId}
 import org.joda.time.format.DateTimeFormat
 import play.api.mvc.{Action, Controller}
-import services.{CompaniesHouseAPI, PagedResults}
+import services.{CompanyAuthService, CompanySearchService, PagedResults}
 import slicks.modules.ReportRepo
 
 import scala.concurrent.{ExecutionContext, Future}
 
-class SearchController @Inject()(companiesHouseAPI: CompaniesHouseAPI, reports: ReportRepo, val appConfig: AppConfig)(implicit ec: ExecutionContext)
+class SearchController @Inject()(
+                                  companySearch: CompanySearchService,
+                                  companyAuth: CompanyAuthService,
+                                  reports: ReportRepo,
+                                  val appConfig: AppConfig)(implicit ec: ExecutionContext)
   extends Controller
     with PageHelper {
 
@@ -48,7 +52,7 @@ class SearchController @Inject()(companiesHouseAPI: CompaniesHouseAPI, reports: 
     val title = "Search for a company"
 
     query match {
-      case Some(q) => companiesHouseAPI.searchCompanies(q, pageNumber.getOrElse(1), itemsPerPage.getOrElse(25)).flatMap { results =>
+      case Some(q) => companySearch.searchCompanies(q, pageNumber.getOrElse(1), itemsPerPage.getOrElse(25)).flatMap { results =>
         val countsF = results.items.map { report =>
           reports.byCompanyNumber(report.company_number).map(rs => (report.company_number, rs.length))
         }
@@ -66,7 +70,7 @@ class SearchController @Inject()(companiesHouseAPI: CompaniesHouseAPI, reports: 
   def company(companiesHouseId: CompaniesHouseId, pageNumber: Option[Int]) = Action.async { implicit request =>
     val pageLink = { i: Int => routes.SearchController.company(companiesHouseId, Some(i)).url }
     val result = for {
-      co <- OptionT(companiesHouseAPI.find(companiesHouseId))
+      co <- OptionT(companySearch.find(companiesHouseId))
       rs <- OptionT.liftF(reports.byCompanyNumber(companiesHouseId).map(rs => PagedResults.page(rs.flatMap(_.filed), pageNumber.getOrElse(1))))
     } yield {
       val searchCrumb = Breadcrumb(routes.SearchController.search(None, None, None), searchForReports)
